### 2020-11-24 #2861 @arfon
@samaygarg - this PR should fix your paper compilation issues: https://github.com/samaygarg/fuelcell/pull/4
