### 2021-01-04 #2940 @whedon
[REVIEW]: fuelcell: A Python package and graphical user interface for electrochemical data analysis
**Submitting author:** @samaygarg (<a href="http://orcid.org/0000-0002-9872-6429">Samay Garg</a>)
**Repository:** <a href="https://github.com/samaygarg/fuelcell" target ="_blank">https://github.com/samaygarg/fuelcell</a>
**Version:** v1.0.1
