### 2021-02-03 03bab37
fixed matplotlib deprecation issues
