+++ fuelcell/datums.py
@@ -853,3 +853,7 @@ def find_steps(arr, threshold=5):
 	diffs = np.abs(np.diff(arr))
 	splits = np.where(diffs > threshold)[0] + 1
 	return splits 
+
+
+
+	
