+++ gui/fuelcell_gui.py
@@ -22,14 +22,15 @@ from PyQt5 import QtCore, QtWidgets, QtGui
 from PyQt5.QtWidgets import *
 from PyQt5.QtGui import QDesktopServices, QFont, QPalette, QColor
 from PyQt5.QtCore import Qt
-
-from matplotlib.backends.qt_compat import is_pyqt5
-if is_pyqt5():
-    from matplotlib.backends.backend_qt5agg import (
-        FigureCanvas, NavigationToolbar2QT as NavigationToolbar)
-else:
-    from matplotlib.backends.backend_qt4agg import (
-        FigureCanvas, NavigationToolbar2QT as NavigationToolbar)
+from PyQt5.QtCore import qVersion
+from matplotlib.backends.backend_qt5agg import (FigureCanvas, NavigationToolbar2QT as NavigationToolbar)
+# from matplotlib.backends.qt_compat import is_pyqt5
+# if qVersion() == 5:
+#     from matplotlib.backends.backend_qt5agg import (
+#         FigureCanvas, NavigationToolbar2QT as NavigationToolbar)
+# else:
+#     from matplotlib.backends.backend_qt4agg import (
+#         FigureCanvas, NavigationToolbar2QT as NavigationToolbar)
 
 import matplotlib
 from matplotlib.figure import Figure
@@ -457,9 +458,9 @@ class FuelcellUI(QTabWidget):
 		self.data_tab = self.makeTab(self.datums_layout(), 'Data Processing')
 		self.visuals_tab = self.makeTab(self.visuals_layout(), 'Visualization')
 		self.tafel_tab = self.makeTab(self.tafel_layout(), 'Tafel Analysis')
-		self.bayes_tab = self.makeTab(self.bayes_layout(), 'Bayesian Tafel Analysis')
+		# self.bayes_tab = self.makeTab(self.bayes_layout(), 'Bayesian Tafel Analysis')
 		self.eis_tab = self.makeTab(self.eis_layout(), 'HFR Analysis')
-		self.datahub_tab = self.makeTab(self.datahub_layout(), 'Datahub Upload')
+		# self.datahub_tab = self.makeTab(self.datahub_layout(), 'Datahub Upload')
 
 		self.data_dict = {}
 		self.eis_dict = {}
@@ -486,22 +487,22 @@ class FuelcellUI(QTabWidget):
 	def set_max_width(self, widget, scale=1):
 		size = widget.sizeHint()
 		w = size.width()
-		widget.setMaximumWidth(w*scale)
+		widget.setMaximumWidth(int(w*scale))
 
 	def set_min_width(self, widget, scale=1):
 		size = widget.sizeHint()
 		w = size.width()
-		widget.setMinimumWidth(w*scale)
+		widget.setMinimumWidth(int(w*scale))
 
 	def set_max_height(self, widget, scale=1):
 		size = widget.sizeHint()
 		h = size.height()
-		widget.setMaximumHeight(h*scale)
+		widget.setMaximumHeight(int(h*scale))
 
 	def set_min_height(self, widget, scale=1):
 		size = widget.sizeHint()
 		h = size.height()
-		widget.setMinimumHeight(h*scale)
+		widget.setMinimumHeight(int(h*scale))
 	
 	def get_all_files(self, dir, valid=None):
 		allfiles = os.listdir(dir)
@@ -1523,7 +1524,7 @@ class FuelcellUI(QTabWidget):
 			self.linestyle_menu_vis.setCurrentText(linestyle)
 			self.linewidth_txtbx_vis.setValue(linewidth)
[…]
