### 2021-02-04 d13b26b
updated gui docs page with new install instructions
