+++ docs/gui.rst
@@ -1,7 +1,7 @@
 Graphical User Interface
 =========================
 
-The GUI for the fuelcell package can be downloaded as a standalone executable file. This GUI can be used on either Mac or Windows operating systems, and the latest version can be downloaded `here <https://github.com/samaygarg/fuelcell/tree/master/gui/fuelcell_gui>`_.
+The GUI for the fuelcell package can be downloaded as a standalone executable file and can be used on either Mac or Windows operating systems. To use the latest version GUI, download the zip file which contains the GUI appropriate to your operating system from the `github releases page <https://github.com/samaygarg/fuelcell/releases>`_. If you are using MacOS, you will likely have to give permission for the program to run after trying to open in for the first time. This can be done by going to System Preferences -> Security & Privacy -> General and clicking "Open Anyway"
 To report any problems that arise while using the GUI, open a `new issue <https://github.com/samaygarg/fuelcell/issues>`_ on github.
 
 New in version 0.4.0
