### 2021-02-04 bddf774
updated gui docs page with new install instructions
