+++ docs/gui.rst
@@ -4,9 +4,6 @@ Graphical User Interface
 The GUI for the fuelcell package can be downloaded as a standalone executable file and can be used on either Mac or Windows operating systems. To use the latest version GUI, download the zip file which contains the GUI appropriate to your operating system from the `github releases page <https://github.com/samaygarg/fuelcell/releases>`_. If you are using MacOS, you will likely have to give permission for the program to run after trying to open in for the first time. This can be done by going to System Preferences -> Security & Privacy -> General and clicking "Open Anyway"
 To report any problems that arise while using the GUI, open a `new issue <https://github.com/samaygarg/fuelcell/issues>`_ on github.
 
-New in version 0.4.0
+New in version 0.5.2
 ---------------------
-* improved line customization abilities
-* bug fixes
-* added support for EIS  data and nyquist plots
-
+* resolved deprecation issues to ensure forward compatibility
