### 2021-02-15 399836c
fixed error handling
