### 2021-03-03 #2940 @jlopata21
fuelcell_gui.py functions as advertised, allowing users to process experimental data efficiently. There are many ways in which it can be improved, e.g. by expanding file format recognition to include raw data files from different potentiostat manufacturers or including some basic 0-D modeling capability. However, these are things that should be developed in time by the community of contributors. Now that the GUI functions properly, we are willing to accept this work as is.
