### 2021-03-03 #2940 @danielskatz
The review thread stays open until the submission is published or rejected.
